+++ server/.env.example
@@ -1,0 +1,2 @@
+PORT=8000
+CORS_ORIGIN='*'
