+++ server/src/index.js
@@ -1,14 +1,15 @@
 import dotenv from "dotenv";
 import { app } from "./app.js";
-import connectDB from "./db/connect.js";
+import connectDB from "./services/db.js";
 dotenv.config({
   path: "./.env",
 });
 
 connectDB()
   .then(() => {
-    app.listen(process.env.PORT || 8000, () => {
-      console.log(`server is running at port ${process.env.PORT}`);
+    const PORT = process.env.PORT || 8000;
+    app.listen(PORT, () => {
+      console.log(`server is running at port ${PORT}`);
     });
   })
   .catch((err) => {
